+++ packages/av-cliper/src/clips/mp4-clip.ts
@@ -2,11 +2,13 @@ import { MP4Info, MP4Sample } from '@webav/mp4box.js';
 import {
   audioResample,
   concatPCMFragments,
+  autoReadStream,
   extractPCM4AudioData,
   sleep,
 } from '../av-utils';
 import { Log } from '../log';
 import { extractFileConfig, sample2ChunkOpts } from '../mp4-utils/mp4box-utils';
+import { SampleTransform } from '../mp4-utils/sample-transform';
 import { DEFAULT_AUDIO_CONF, IClip } from './iclip';
 import { file, tmpfile, write } from 'opfs-tools';
 import { fileToMP4Samples } from '../mp4-utils/opfs-2-sample';
@@ -36,7 +38,11 @@ interface MP4ClipOpts {
   __unsafe_hardwareAcceleration__?: HardwarePreference;
 }
 
-type ExtMP4Sample = Omit<MP4Sample, 'data'> & { deleted?: boolean; data: null };
+type ExtMP4Sample = Omit<MP4Sample, 'data'> & {
+  is_idr: boolean;
+  deleted?: boolean;
+  data: null | Uint8Array;
+};
 
 type LocalFileReader = Awaited<ReturnType<OPFSToolFile['createReader']>>;
 
@@ -144,7 +150,7 @@ export class MP4Clip implements IClip {
       this.#videoSamples = videoSamples;
       this.#audioSamples = audioSamples;
       this.#decoderConf = decoderConf;
-      const { videoFrameFinder, audioFrameFinder } = genDeocder(
+      const { videoFrameFinder, audioFrameFinder } = genDecoder(
         {
           video:
             decoderConf.video == null
@@ -198,42 +204,10 @@ export class MP4Clip implements IClip {
       });
     }
 
-    let audioReady = false;
-    let videoReady = false;
-    let timeoutTimer = 0;
-    const [audio, video] = (await Promise.race([
-      Promise.all([
-        this.#audioFrameFinder?.find(time).then((rs) => {
-          audioReady = true;
-          return rs;
-        }) ?? [],
-        this.#videoFrameFinder?.find(time).then((rs) => {
-          videoReady = true;
-          return rs;
-        }),
-      ]),
-      new Promise<[]>((_, reject) => {
-        timeoutTimer = self.setTimeout(() => {
-          let errMsg = `MP4Clip.tick timeout, ${JSON.stringify({
-            videoReady,
-            audioReady,
-          })}, `;
-
-          const aFinder = this.#audioFrameFinder;
-          if (!audioReady && aFinder != null) {
-            errMsg += JSON.stringify(aFinder.getState());
-            aFinder?.reset();
-          }
-          const vFinder = this.#videoFrameFinder;
-          if (!videoReady && vFinder != null) {
-            errMsg += JSON.stringify(vFinder.getState());
-            vFinder.reset();
-          }
-          reject(Error(errMsg));
-        }, 3000);
-      }),
-    ])) as [Float32Array[], VideoFrame];
-    clearTimeout(timeoutTimer);
+    const [audio, video] = await Promise.all([
+      this.#audioFrameFinder?.find(time) ?? [],
+      this.#videoFrameFinder?.find(time),
+    ]);
 
     if (video == null) {
       return await this.tickInterceptor(time, {
@@ -251,9 +225,9 @@ export class MP4Clip implements IClip {
 
   #thumbAborter = new AbortController();
   /**
-   * Generate thumbnails, default generate 100px width thumbnails by every key frame.
+   * 生成缩略图，默认每个关键帧生成一个 100px 宽度的缩略图。
    *
-   * @param imgWidth thumbnail width, default 100
+   * @param imgWidth 缩略图宽度，默认 100
    * @param opts Partial<ThumbnailOpts>
    * @returns Promise<Array<{ ts: number; img: Blob }>>
    */
@@ -491,7 +465,7 @@ function genMeta(
   return meta;
 }
 
-function genDeocder(
+function genDecoder(
   decoderConf: MP4DecoderConf,
   localFileReader: LocalFileReader,
   videoSamples: ExtMP4Sample[],
@@ -586,6 +560,7 @@ async function parseMP4Stream(localFile: OPFSToolFile, opts: MP4ClipOpts = {}) {
         // 修复首帧黑帧
         const firstSample = videoSamples[0];
         if (firstSample != null && firstSample.cts < 200e3) {
+          firstSample.duration += firstSample.cts;
           firstSample.cts = 0;
         }
         Log.info('mp4 stream parsed');
@@ -598,14 +573,23 @@ async function parseMP4Stream(localFile: OPFSToolFile, opts: MP4ClipOpts = {}) {
     });
   });
 
-  function normalizeTimescale(s: MP4Sample, delta = 0) {
+  function normalizeTimescale(
+    s: MP4Sample,
+    delta = 0,
+    sampleType: 'video' | 'audio',
+  ) {
     return {
       ...s,
+      is_idr:
+        sampleType === 'video' &&
+        s.is_sync &&
+        isIDRFrame(s.data, s.description.type),
       cts: ((s.cts - delta) / s.timescale) * 1e6,
       dts: ((s.dts - delta) / s.timescale) * 1e6,
       duration: (s.duration / s.timescale) * 1e6,
       timescale: 1e6,
-      data: null,
+      // 音频数据量可控，直接保存在内存中
+      data: sampleType === 'video' ? null : s.data,
     };
   }
 }
@@ -619,16 +603,16 @@ class VideoFrameFinder {
   ) {}
 
   #ts = 0;
-  #curAborter = { abort: false };
+  #curAborter = { abort: false, st: performance.now() };
   find = async (time: number): Promise<VideoFrame | null> => {
     if (this.#dec == null || time <= this.#ts || time - this.#ts > 3e6) {
-      this.reset();
+      this.#reset(time);
     }
 
     this.#curAborter.abort = true;
     this.#ts = time;
 
-    this.#curAborter = { abort: false };
+    this.#curAborter = { abort: false, st: performance.now() };
     return await this.#parseFrame(time, this.#dec, this.#curAborter);
   };
 
@@ -643,7 +627,7 @@ class VideoFrameFinder {
   #parseFrame = async (
     time: number,
     dec: VideoDecoder | null,
-    aborter: { abort: boolean },
+    aborter: { abort: boolean; st: number },
   ): Promise<VideoFrame | null> => {
     if (dec == null || dec.state === 'closed' || aborter.abort) return null;
 
@@ -655,76 +639,119 @@ class VideoFrameFinder {
       // 第一帧过期，找下一帧
       if (time > vf.timestamp + (vf.duration ?? 0)) {
         vf.close();
-        return this.#parseFrame(time, dec, aborter);
+        return await this.#parseFrame(time, dec, aborter);
+      }
+
+      if (this.#videoFrames.length < 10) {
+        // 预解码 避免等待
+        this.#startDecode(dec).catch((err) => {
+          this.#reset(time);
+          throw err;
+        });
       }
       // 符合期望
       return vf;
     }
 
     // 缺少帧数据
-    if (this.#outputFrameCnt < this.#inputChunkCnt && dec.decodeQueueSize > 0) {
+    if (
+      this.#decoding ||
+      (this.#outputFrameCnt < this.#inputChunkCnt && dec.decodeQueueSize > 0)
+    ) {
+      if (performance.now() - aborter.st > 3e3) {
+        throw Error(
+          `MP4Clip.tick video timeout, ${JSON.stringify(this.#getState())}`,
+        );
+      }
       // 解码中，等待，然后重试
       await sleep(15);
     } else if (this.#videoDecCusorIdx >= this.samples.length) {
       // decode completed
       return null;
     } else {
-      // 启动解码任务，然后重试
-      let endIdx = this.#videoDecCusorIdx + 1;
-      // 该 GoP 时间区间有时间匹配，且未被删除的帧
-      let hasValidFrame = false;
-      for (; endIdx < this.samples.length; endIdx++) {
-        const s = this.samples[endIdx];
-        if (!hasValidFrame && !s.deleted && time < s.cts + s.duration) {
-          hasValidFrame = true;
-        }
-        // 找一个 GoP，所以是下一个关键帧结束
-        if (s.is_sync) break;
+      try {
+        await this.#startDecode(dec);
+      } catch (err) {
+        this.#reset(time);
+        throw err;
       }
-
-      if (hasValidFrame) {
-        const samples = this.samples.slice(this.#videoDecCusorIdx, endIdx);
-        if (samples[0]?.is_sync !== true) {
-          Log.warn('First sample not key frame');
-        } else {
-          const chunks = await Promise.all(
-            samples.map((s) =>
-              sample2Chunk(s, EncodedVideoChunk, this.localFileReader),
-            ),
-          );
-          // Wait for the previous asynchronous operation to complete, at which point the task may have already been terminated
-          if (aborter.abort) return null;
-
-          this.#lastVfDur = chunks[0]?.duration ?? 0;
-          decodeGoP(dec, chunks, {
-            onDecodingError: (err) => {
-              if (this.#downgradeSoftDecode) {
-                throw err;
-              } else {
-                this.#downgradeSoftDecode = true;
-                Log.warn('Downgrade to software decode');
-                this.reset();
-              }
-            },
-          });
-
-          this.#inputChunkCnt += chunks.length;
-        }
-      }
-      this.#videoDecCusorIdx = endIdx;
     }
-    return this.#parseFrame(time, dec, aborter);
+    return await this.#parseFrame(time, dec, aborter);
   };
 
-  reset = () => {
+  #decoding = false;
+  #startDecode = async (dec: VideoDecoder) => {
+    if (this.#decoding) return;
+    this.#decoding = true;
+
+    // 启动解码任务，然后重试
+    let endIdx = this.#videoDecCusorIdx + 1;
+    // 该 GoP 时间区间有时间匹配，且未被删除的帧
+    let hasValidFrame = false;
+    for (; endIdx < this.samples.length; endIdx++) {
+      const s = this.samples[endIdx];
+      if (!hasValidFrame && !s.deleted) {
+        hasValidFrame = true;
+      }
+      // 找一个 GoP，所以是下一个 IDR 帧结束
+      if (s.is_idr) break;
+    }
+
+    if (hasValidFrame) {
+      const samples = this.samples.slice(this.#videoDecCusorIdx, endIdx);
+      if (samples[0]?.is_idr !== true) {
+        Log.warn('First sample not idr frame');
+      } else {
+        const chunks = await videosamples2Chunks(samples, this.localFileReader);
+        // Wait for the previous asynchronous operation to complete, at which point the task may have already been terminated
+        if (dec.state === 'closed') return;
+
+        this.#lastVfDur = chunks[0]?.duration ?? 0;
+        decodeGoP(dec, chunks, {
+          onDecodingError: (err) => {
+            if (this.#downgradeSoftDecode) {
+              throw err;
+            } else {
+              this.#downgradeSoftDecode = true;
+              Log.warn('Downgrade to software decode');
+              this.#reset();
+            }
+          },
+        });
+
+        this.#inputChunkCnt += chunks.length;
+      }
+    }
+    this.#videoDecCusorIdx = endIdx;
+    this.#decoding = false;
+  };
+
+  #reset = (time?: number) => {
+    this.#decoding = false;
     this.#videoFrames.forEach((f) => f.close());
     this.#videoFrames = [];
-    this.#videoDecCusorIdx = 0;
+    if (time == null || time === 0) {
+      this.#videoDecCusorIdx = 0;
+    } else {
+      let keyIdx = 0;
+      for (let i = 0; i < this.samples.length; i++) {
+        const s = this.samples[i];
+        if (s.is_idr) keyIdx = i;
+        if (s.cts < time) continue;
+        this.#videoDecCusorIdx = keyIdx;
+        break;
+      }
+    }
     this.#inputChunkCnt = 0;
     this.#outputFrameCnt = 0;
     if (this.#dec?.state !== 'closed') this.#dec?.close();
     this.#dec = new VideoDecoder({
       output: (vf) => {
+        this.#outputFrameCnt += 1;
+        if (vf.timestamp === -1) {
+          vf.close();
+          return;
+        }
         let rsVf = vf;
         if (vf.duration == null) {
           rsVf = new VideoFrame(vf, {
@@ -732,10 +759,11 @@ class VideoFrameFinder {
           });
           vf.close();
         }
-        this.#outputFrameCnt += 1;
         this.#videoFrames.push(rsVf);
       },
-      error: Log.error,
+      error: (err) => {
+        Log.error(`MP4Clip VideoDecoder err: ${err.message}`);
+      },
     });
     this.#dec.configure({
       ...this.conf,
@@ -745,7 +773,7 @@ class VideoFrameFinder {
     });
   };
 
-  getState = () => ({
+  #getState = () => ({
     time: this.#ts,
     decState: this.#dec?.state,
     decQSize: this.#dec?.decodeQueueSize,
@@ -781,11 +809,11 @@ class AudioFrameFinder {
   }
 
   #dec: ReturnType<typeof createAudioChunksDecoder> | null = null;
-  #curAborter = { abort: false };
+  #curAborter = { abort: false, st: performance.now() };
   find = async (time: number): Promise<Float32Array[]> => {
     // 前后获取音频数据差异不能超过 100ms
     if (this.#dec == null || time <= this.#ts || time - this.#ts > 0.1e6) {
-      this.reset();
+      this.#reset();
       this.#ts = time;
       for (let i = 0; i < this.samples.length; i++) {
         if (this.samples[i].cts < time) continue;
@@ -799,183 +827,172 @@ class AudioFrameFinder {
     const deltaTime = time - this.#ts;
     this.#ts = time;
 
-    this.#curAborter = { abort: false };
+    this.#curAborter = { abort: false, st: performance.now() };
     return await this.#parseFrame(deltaTime, this.#dec, this.#curAborter);
   };
 
   #ts = 0;
   #decCusorIdx = 0;
-  #decoding = false;
-  #pcmData: [Float32Array, Float32Array] = [
-    new Float32Array(0), // left chan
-    new Float32Array(0), // right chan
-  ];
+  #pcmData: {
+    frameCnt: number;
+    data: [Float32Array, Float32Array][];
+  } = {
+    frameCnt: 0,
+    data: [],
+  };
   #parseFrame = async (
     deltaTime: number,
     dec: ReturnType<typeof createAudioChunksDecoder> | null = null,
-    aborter: { abort: boolean },
+    aborter: { abort: boolean; st: number },
   ): Promise<Float32Array[]> => {
     if (dec == null || aborter.abort || dec.state === 'closed') return [];
 
-    const frameCnt = Math.ceil(deltaTime * (this.#sampleRate / 1e6));
-    if (frameCnt === 0) return [];
+    const emitFrameCnt = Math.ceil(deltaTime * (this.#sampleRate / 1e6));
+    if (emitFrameCnt === 0) return [];
 
     // 数据满足需要
-    if (this.#pcmData[0].length > frameCnt) {
-      const audio = [
-        this.#pcmData[0].slice(0, frameCnt),
-        this.#pcmData[1].slice(0, frameCnt),
-      ];
-      this.#pcmData[0] = this.#pcmData[0].slice(frameCnt);
-      this.#pcmData[1] = this.#pcmData[1].slice(frameCnt);
-      return audio;
+    const ramainFrameCnt = this.#pcmData.frameCnt - emitFrameCnt;
+    if (ramainFrameCnt > 0) {
+      // 剩余音频数据小于 100ms
+      if (ramainFrameCnt < DEFAULT_AUDIO_CONF.sampleRate / 10) {
+        this.#startDecode(dec);
+      }
+      return emitAudioFrames(this.#pcmData, emitFrameCnt);
     }
 
-    if (this.#decoding) {
+    if (dec.decodeQueueSize > 10) {
+      if (performance.now() - aborter.st > 3e3) {
+        aborter.abort = true;
+        throw Error(
+          `MP4Clip.tick audio timeout, ${JSON.stringify(this.#getState())}`,
+        );
+      }
       // 解码中，等待
       await sleep(15);
     } else if (this.#decCusorIdx >= this.samples.length - 1) {
       // decode completed
       return [];
     } else {
-      // 启动解码任务
-      const samples = [];
-      for (let i = this.#decCusorIdx; i < this.samples.length; i++) {
-        this.#decCusorIdx = i;
-        const s = this.samples[i];
-        if (s.deleted) continue;
-        if (samples.length >= 10) break;
-        samples.push(s);
-      }
-
-      this.#decoding = true;
-      dec.decode(
-        await Promise.all(
-          samples.map((s) =>
-            sample2Chunk(s, EncodedAudioChunk, this.localFileReader),
-          ),
-        ),
-        (pcmArr, done) => {
-          if (pcmArr.length === 0) return;
-          // 音量调节
-          if (this.#volume !== 1) {
-            for (const pcm of pcmArr)
-              for (let i = 0; i < pcm.length; i++) pcm[i] *= this.#volume;
-          }
-          // 补齐双声道
-          if (pcmArr.length === 1) pcmArr = [pcmArr[0], pcmArr[0]];
-
-          this.#pcmData = concatPCMFragments([this.#pcmData, pcmArr]) as [
-            Float32Array,
-            Float32Array,
-          ];
-          if (done) this.#decoding = false;
-        },
-      );
+      this.#startDecode(dec);
     }
     return this.#parseFrame(deltaTime, dec, aborter);
   };
 
-  reset = () => {
-    this.#ts = 0;
-    this.#decCusorIdx = 0;
-    this.#pcmData = [
-      new Float32Array(0), // left chan
-      new Float32Array(0), // right chan
-    ];
-    this.#dec?.close();
-    this.#decoding = false;
-    this.#dec = createAudioChunksDecoder(
-      this.conf,
-      DEFAULT_AUDIO_CONF.sampleRate,
+  #startDecode = (dec: ReturnType<typeof createAudioChunksDecoder>) => {
+    if (dec.decodeQueueSize > 100) return;
+    // 启动解码任务
+    const samples = [];
+    let i = this.#decCusorIdx;
+    while (i < this.samples.length) {
+      const s = this.samples[i];
+      i += 1;
+      if (s.deleted) continue;
+      samples.push(s);
+      if (samples.length >= 10) break;
+    }
+    this.#decCusorIdx = i;
+
+    dec.decode(
+      samples.map(
+        (s) =>
+          new EncodedAudioChunk({
+            type: 'key',
+            timestamp: s.cts,
+            duration: s.duration,
+            data: s.data!,
+          }),
+      ),
     );
   };
 
-  getState = () => ({
+  #reset = () => {
+    this.#ts = 0;
+    this.#decCusorIdx = 0;
+    this.#pcmData = {
+      frameCnt: 0,
+      data: [],
+    };
+    this.#dec?.close();
+    this.#dec = createAudioChunksDecoder(
+      this.conf,
+      {
+        resampleRate: DEFAULT_AUDIO_CONF.sampleRate,
+        volume: this.#volume,
+      },
+      (pcmArr) => {
+        this.#pcmData.data.push(pcmArr as [Float32Array, Float32Array]);
+        this.#pcmData.frameCnt += pcmArr[0].length;
+      },
+    );
+  };
+
+  #getState = () => ({
     time: this.#ts,
     decState: this.#dec?.state,
-    decoding: this.#decoding,
     decQSize: this.#dec?.decodeQueueSize,
     decCusorIdx: this.#decCusorIdx,
     sampleLen: this.samples.length,
-    pcmLen: this.#pcmData[0]?.length,
+    pcmLen: this.#pcmData.frameCnt,
   });
 
   destroy = () => {
     this.#dec = null;
     this.#curAborter.abort = true;
-    this.#pcmData = [
-      new Float32Array(0), // left chan
-      new Float32Array(0), // right chan
-    ];
+    this.#pcmData = {
+      frameCnt: 0,
+      data: [],
+    };
     this.localFileReader.close();
   };
 }
 
 function createAudioChunksDecoder(
   decoderConf: AudioDecoderConfig,
-  resampleRate: number,
+  opts: { resampleRate: number; volume: number },
+  outputCb: (pcm: Float32Array[]) => void,
 ) {
-  type OutputHandle = (pcm: Float32Array[], done: boolean) => void;
+  const outputHandler = (pcmArr: Float32Array[]) => {
+    if (pcmArr.length === 0) return;
+    // 音量调节
+    if (opts.volume !== 1) {
+      for (const pcm of pcmArr)
+        for (let i = 0; i < pcm.length; i++) pcm[i] *= opts.volume;
+    }
 
-  let curCb: ((pcm: Float32Array[], done: boolean) => void) | null = null;
-  const needResample = resampleRate !== decoderConf.sampleRate;
-  const resampleQ = createPromiseQueue<[Float32Array[], boolean]>(
-    ([resampedPCM, done]) => {
-      curCb?.(resampedPCM, done);
-    },
-  );
+    // 补齐双声道
+    if (pcmArr.length === 1) pcmArr = [pcmArr[0], pcmArr[0]];
 
+    outputCb(pcmArr);
+  };
+  const resampleQ = createPromiseQueue<Float32Array[]>(outputHandler);
+
+  const needResample = opts.resampleRate !== decoderConf.sampleRate;
   const adec = new AudioDecoder({
     output: (ad) => {
       const pcm = extractPCM4AudioData(ad);
-      const done = adec.decodeQueueSize === 0;
       if (needResample) {
-        resampleQ(async () => [
-          await audioResample(pcm, ad.sampleRate, {
-            rate: resampleRate,
+        resampleQ(() =>
+          audioResample(pcm, ad.sampleRate, {
+            rate: opts.resampleRate,
             chanCount: ad.numberOfChannels,
           }),
-          done,
-        ]);
+        );
       } else {
-        curCb?.(pcm, done);
+        outputHandler(pcm);
       }
       ad.close();
     },
-    error: Log.error,
+    error: (err) => {
+      Log.error(`MP4Clip AudioDecoder err: ${err.message}`);
+    },
   });
   adec.configure(decoderConf);
 
-  let tasks: Array<{ chunks: EncodedAudioChunk[]; cb: OutputHandle }> = [];
-  async function run() {
-    if (curCb != null || adec.state !== 'configured') return;
-
-    const t = tasks.shift();
-    if (t == null) return;
-    if (t.chunks.length <= 0) {
-      t.cb([], true);
-      run().catch(Log.error);
-      return;
-    }
-
-    curCb = (pcm, done) => {
-      t.cb(pcm, done);
-      if (done) {
-        curCb = null;
-        run().catch(Log.error);
-      }
-    };
-    for (const chunk of t.chunks) adec.decode(chunk);
-  }
-
   return {
-    decode(chunks: EncodedAudioChunk[], cb: OutputHandle) {
-      tasks.push({ chunks, cb });
-      run().catch(Log.error);
+    decode(chunks: EncodedAudioChunk[]) {
+      for (const chunk of chunks) adec.decode(chunk);
     },
     close() {
-      curCb = null;
       if (adec.state !== 'closed') adec.close();
     },
     get state() {
@@ -1016,22 +1033,74 @@ function createPromiseQueue<T extends any>(onResult: (data: T) => void) {
   };
 }
 
-type Constructor<T> = {
-  new (...args: any[]): T;
-};
+function emitAudioFrames(
+  pcmData: { frameCnt: number; data: [Float32Array, Float32Array][] },
+  emitCnt: number,
+) {
+  // todo: perf 重复利用内存空间
+  const audio = [new Float32Array(emitCnt), new Float32Array(emitCnt)];
+  let offset = 0;
+  let i = 0;
+  for (; i < pcmData.data.length; ) {
+    const [chan0, chan1] = pcmData.data[i];
+    if (offset + chan0.length > emitCnt) {
+      const gapCnt = emitCnt - offset;
+      audio[0].set(chan0.subarray(0, gapCnt), offset);
+      audio[1].set(chan1.subarray(0, gapCnt), offset);
+      pcmData.data[i][0] = chan0.subarray(gapCnt, chan0.length);
+      pcmData.data[i][1] = chan1.subarray(gapCnt, chan1.length);
+      break;
+    } else {
+      audio[0].set(chan0, offset);
+      audio[1].set(chan1, offset);
+      offset += chan0.length;
+      i++;
+    }
+  }
+  pcmData.data = pcmData.data.slice(i);
+  pcmData.frameCnt -= emitCnt;
+  return audio;
+}
 
-async function sample2Chunk<T extends EncodedAudioChunk | EncodedVideoChunk>(
-  s: ExtMP4Sample,
-  clazz: Constructor<T>,
+async function videosamples2Chunks(
+  samples: ExtMP4Sample[],
   reader: Awaited<ReturnType<OPFSToolFile['createReader']>>,
-): Promise<T> {
-  // todo: perf
-  const data = await reader.read(s.size, { at: s.offset });
-  return new clazz(
-    // todo: perf
-    sample2ChunkOpts({
-      ...s,
-      data,
+): Promise<EncodedVideoChunk[]> {
+  const first = samples[0];
+  const last = samples.at(-1);
+  if (last == null) return [];
+
+  const rangSize = last.offset + last.size - first.offset;
+  if (rangSize < 30e6) {
+    // 单次读取数据小于 30M，就一次性读取数据，降低 IO 频次
+    const data = new Uint8Array(
+      await reader.read(rangSize, { at: first.offset }),
+    );
+    return samples.map((s) => {
+      const offset = s.offset - first.offset;
+      let sData = data.subarray(offset, offset + s.size);
+      if (s.is_idr) sData = removeSEIForIDR(sData);
+      return new EncodedVideoChunk({
+        type: s.is_sync ? 'key' : 'delta',
+        timestamp: s.cts,
+        duration: s.duration,
+        data: sData,
+      });
+    });
+  }
+
+  return await Promise.all(
+    samples.map(async (s) => {
+      let sData = await reader.read(s.size, {
+        at: s.offset,
+      });
+      if (s.is_idr) sData = removeSEIForIDR(new Uint8Array(sData));
+      return new EncodedVideoChunk({
+        type: s.is_sync ? 'key' : 'delta',
+        timestamp: s.cts,
+        duration: s.duration,
+        data: sData,
+      });
     }),
   );
 }
@@ -1060,7 +1129,7 @@ function splitVideoSampleByTime(videoSamples: ExtMP4Sample[], time: number) {
   for (let i = 0; i < videoSamples.length; i++) {
     const s = videoSamples[i];
     if (hitIdx === -1 && time < s.cts) hitIdx = i - 1;
-    if (s.is_sync) {
+    if (s.is_idr) {
       if (hitIdx === -1) {
         gopStartIdx = i;
       } else {
@@ -1085,14 +1154,14 @@ function splitVideoSampleByTime(videoSamples: ExtMP4Sample[], time: number) {
   }
 
   const postSlice = videoSamples
-    .slice(hitSample.is_sync ? gopEndIdx : gopStartIdx)
+    .slice(hitSample.is_idr ? gopEndIdx : gopStartIdx)
     .map((s) => ({ ...s, cts: s.cts - time }));
-  for (let i = 0; i < gopEndIdx - gopEndIdx; i++) {
-    const s = preSlice[i];
-    if (s.cts < 0) {
-      s.deleted = true;
-      s.cts = -1;
-    }
+
+  for (const s of postSlice) {
+    if (s.cts >= 0) break;
+    // 将第一个 GoP 中前半部分标记为 deleted
+    s.deleted = true;
+    s.cts = -1;
   }
 
   return [preSlice, postSlice];
@@ -1120,34 +1189,12 @@ function decodeGoP(
   dec: VideoDecoder,
   chunks: EncodedVideoChunk[],
   opts: {
-    idrFrameDowngrade?: boolean;
     onDecodingError?: (err: Error) => void;
   },
 ) {
   let i = 0;
-  try {
-    if (dec.state !== 'configured') return;
-    for (; i < chunks.length; i++) dec.decode(chunks[i]);
-  } catch (err) {
-    if (opts.idrFrameDowngrade || !(err instanceof Error)) throw err;
-    if (
-      i === 0 &&
-      err.message.includes('A key frame is required after configure')
-    ) {
-      // 第一帧携带 SEI 信息会导致解码失败
-      const newChunk = removeNonIDRData(chunks[0]);
-      if (newChunk == null) throw err;
-
-      Log.warn('remove non IDR data, retry decode');
-      chunks[0] = newChunk;
-      decodeGoP(dec, chunks, {
-        ...opts,
-        idrFrameDowngrade: true,
-      });
-    } else {
-      throw err;
-    }
-  }
+  if (dec.state !== 'configured') return;
+  for (; i < chunks.length; i++) dec.decode(chunks[i]);
 
   // windows 某些设备 flush 可能不会被 resolved，所以不能 await flush
   dec.flush().catch((err) => {
@@ -1167,26 +1214,29 @@ function decodeGoP(
 }
 
 // 当 IDR 帧前面携带其它数据（如 SEI）可能导致解码失败
-function removeNonIDRData(chunk: EncodedVideoChunk) {
-  const buf = new ArrayBuffer(chunk.byteLength);
-  chunk.copyTo(buf);
-  const u8 = new Uint8Array(buf);
-  let i = 0;
-  for (; i < chunk.byteLength - 4; ) {
-    if ((u8[i + 4] & 0x1f) === 5) break;
-    // 跳至下一个 NALU 继续检查
-    i += (u8[i] << 24) + (u8[i + 1] << 16) + (u8[i + 2] << 8) + u8[i + 3] + 4;
+function removeSEIForIDR(u8buf: Uint8Array) {
+  const dv = new DataView(u8buf.buffer, u8buf.byteOffset, u8buf.byteLength);
+  if ((dv.getUint8(4) & 0x1f) === 6) {
+    return u8buf.subarray(dv.getUint32(0) + 4);
   }
+  return u8buf;
+}
 
-  if (i < buf.byteLength) {
-    return new EncodedVideoChunk({
-      type: chunk.type,
-      timestamp: chunk.timestamp,
-      duration: chunk.duration ?? 0,
-      data: buf.slice(i),
-    });
+function isIDRFrame(u8Arr: Uint8Array, type: MP4Sample['description']['type']) {
+  if (type !== 'avc1' && type !== 'hvc1') return false;
+
+  const dv = new DataView(u8Arr.buffer);
+  let i = 0;
+  for (; i < u8Arr.byteLength - 4; ) {
+    if (type === 'avc1') {
+      if ((dv.getUint8(i + 4) & 0x1f) === 5) return true;
+    } else if (type === 'hvc1') {
+      if (((dv.getUint8(i + 4) >> 1) & 0x3f) === 20) return true;
+    }
+    // 跳至下一个 NALU 继续检查
+    i += dv.getUint32(i) + 4;
   }
-  return null;
+  return false;
 }
 
 async function thumbnailByKeyFrame(
@@ -1213,13 +1263,12 @@ async function thumbnailByKeyFrame(
     dec.close();
   });
 
-  const chunks = await Promise.all(
-    samples
-      .filter(
-        (s) =>
-          !s.deleted && s.is_sync && s.cts >= time.start && s.cts <= time.end,
-      )
-      .map((s) => sample2Chunk(s, EncodedVideoChunk, fileReader)),
+  const chunks = await videosamples2Chunks(
+    samples.filter(
+      (s) =>
+        !s.deleted && s.is_sync && s.cts >= time.start && s.cts <= time.end,
+    ),
+    fileReader,
   );
   if (chunks.length === 0 || abortSingl.aborted) return;